requirements.txt
streamlit
pandas
plotly
prophet
tefas-crawler
watchdog